kennzahlenrechner.py
import colorful
import yaml


class Main():

    def __init__(self):
        try:
            with open("./std_kennzahlen.yml", 'r') as stream:
                self.std_kennzahlen_yaml = yaml.load(stream)
            with open("./bilanz.yml", 'r') as stream:
                print("fff")
                self.bilanz_yaml = yaml.load(stream)
        except Exception as e:
            print(colorful.bold_red(e))


if __name__ == "__main__":
    programm = Main()
